Avance de entidades y un servicio

+++ src/main/java/com/negelec/app/productos/repository/IProductosDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.negelec.app.productos.entity.Producto;
 
-public interface IProductosDAO extends JpaRepository<Producto,Long>{
+public interface IProductosDAO extends JpaRepository<Producto,Integer>{
 
 }
